Core: Use ro.build.date to signal mIsUpgrade
M: We use a static fingerprint that is only changed when a new OEM build is released, so every flash shows Android is starting instead of upgrading.  This will fix that.

N: even though we dont have the dexopt sceen on N, this is still needed to delete the correct caches, and grant/deny specific runtime permissions like a true oem update would do

Updated for Nougat By: BeansTown106

Change-Id: I12f95eba4644a45f57d26af90e331eda9f761988
Signed-off-by: spezi77 <spezi77@gmx.com>