Expose and tint settings dashboard icon tint
Doing some research, I found that settings icons are tinted by using
colorControlNormal. While using our dark theme, we define colorControlNormal
and call on the system accent. Everything looks great using the dark theme.

The problem here was that the system was not defining it in the styles while
using the light(default) theme so the icon tint default to #737373 so when
you applied an accent using the light theme, those icons were not being changed.

This commit defines colorControlNormal for the Settings styles and also exposes
it so that we can change this with the accents, all while keeping the stock look
in place. Win win :-)

Change-Id: Ibfd150c31cc4404e95c5cab49e9ddf90124a9b14
Signed-off-by: spezi77 <spezi7713@gmx.net>