Fix StatusBar icons tinting when in split screen
Symptom:
The color of Wifi and Mobile icons doesn't turn white when enter
split-screen mode in landscape, even if color of status bar is dark.

Solution:
Applied every tints to StatusBar icons at onDarkChanged even if they
are not in the area.
Consider the area to the darkIntensity of the argument at onDarkChanged.
And also fixed PhoneWindowManager to send correct bounds to SystemUI
when in split-screen mode.

Bug: 111420765
Change-Id: I3db264a5b8e3e72441f6efd3e20a5e4671affe22
Signed-off-by: spezi77 <spezi7713@gmx.net>