base: FontService: Add a error for non-supported font packages
With the addition of subs fonts, there are some themes that doesn't have a fonts.xml, hence the process of processing fonts making the system crash due to a NPE.
This doesn't fix it. Instead, let's show up a toast informing the user why the font doesn't show up on his selector and patching the NPE, so users doesn't suffer on the system dying when they install a subs theme.

Big thanks to bigrushdog for the help and research on this situation. It's not the fix we wanted but it's the fix we need for now.

Change-Id: I0055549a96a91db97e6e6ecb1ae0d948c5c483ee
Signed-off-by: spezi77 <spezi7713@gmx.net>