LatinIME: fix navigation bar for expanded desktop
When using expanded desktop with hidden navigation bar the
keyboard still sits up where the navbar should be making a
blank area so make the keyboard go down to the bottom of the
screen.

This only helps when using global expanded desktop and not the
per-app options. There's probably a better way to do this, but
this will work for now.

Change-Id: Ie323b5b51a97fddf1a24b478952b23a620f26efe
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>